feature-locking http-api-common to fix wasm build
